Invincible Powerup, Confuser Enemy and Confused Status added
-Invincible Powerup added (grants immunity from hazard volumes and confused status effect temporarily)
-Confuser Enemy added (When close to the player, will follow them and try to shoot them)
-Confuse Orb added (Shoots from Confuser Enemy)
-Confused Status added (Random force and rotation applied to player)

+++ Assets/Scripts/HazardVolume.cs
@@ -4,11 +4,14 @@ using UnityEngine;
 
 public class HazardVolume : MonoBehaviour
 {
-    private void OnTriggerEnter(Collider other)
+    private void OnTriggerStay(Collider other)
     {
         //if collides with player ship, kill it
         PlayerShip playerShip = other.GetComponent<PlayerShip>();
-        if (playerShip != null)
+        if (playerShip != null && !playerShip.invinciblePowerupActive)
+        {
             playerShip.Kill();
+            DelayHelper.DelayAction(this, playerShip.Respawn, 2f);
+        }
     }
 }
